requirements.txt
Django==1.8.2
Pillow==2.8.2
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
colorthief==0.1
djangorestframework==3.1.3
kombu==3.0.26
pytz==2015.4
webcolors==1.5
